MyBook/requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
psycopg2-binary==2.9.1
cassandra-driver==3.29.0
confluent-kafka==1.8.2
pydantic==1.8.2
pytest==6.2.5
pytest-asyncio==0.15.1